+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2021-03-15
+* Added `st2chatops` support and service startup script. (#206)
+
 ## 2020-11-05
 * Deprecate st2resultstracker which is obsolete since the Mistral deprecation with st2 `v3.3.0`.